requirements.txt
# LLM4 - Kaggle P100 / local
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
trl>=0.25.0
datasets>=2.14.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
PyYAML>=6.0
